paths refactor
